Rename utils to util and remove math.h redundant dependency
